"Commit refactoring of addclass component for cleaner code and routing back to studio homepage"
